Added compiler tests for macro invocations
`compiletest_rs` is used to test the error messages emitted from macro
invocations.
Four tests have been added for invocations with incorrect syntax. These
are
 * an empty invocation
 * just a binary operator
 * just a prefix operator
 * just a postfix operator
The tests are all written literately, using the extra folders feature
that was added to `cargo lit`.
Added the `tests` folder to .gitignore, as it is a literate product.