Changed raw pattern parsing to use Syn
Rather than trying to guess at when users are using patterns, we now
instead rely on the `syn` crate's implementation of a parser for
possible Rust patterns. This is currently too powerful, as `syn` allows
for aliases in the raw, which we want to instead capture as a custom
operator - this breaks two tests.
Otherwise, the logic is much more sound, and should report some nicer
bad patterns rather than accepting them silently and ruining everything.