Added type annotation of patterns
This solves the issue of needing type information of a pattern, and
prevents the hardcoding that I previously had.
Patterns can now be annotated with an arrow type Foo -> Bar, such that
the pattern accepts feeds of Foo, and returns results of Bar.
Every top-level pattern requires a type annotation, or the function
can't be generated.
Added a test for a top-level pattern having no type annotation.