Changed type signature to || : a -> b form
The brackets have been reserved to make future changes easier on myself.
It is no longer possible to type-annotate arbitrary pattern parts, but
this should not be missed, since it did absolutely nothing.
All tests have been updated.
Added a dependency on `if_chain`, but this should be cleaned up.